requirements.txt
alembic~=1.14.1
fastapi~=0.115.8
python-dotenv~=1.0.1
sqlalchemy~=2.0.37
fastapi-users[sqlalchemy]
fastapi[all]
uvicorn~=0.34.0
fastapi-cache2[redis]
redis~=5.2.1
gunicorn~=23.0.0
pydantic~=2.10.6
starlette~=0.45.3
aiosqlite~=0.22.1
fakeredis[lua]
pytest~=8.4.2
pytest-asyncio~=1.2.0
coverage~=7.10.7
locust~=2.34.0
